+++ engine/src/main/java/org/destinationsol/SolApplication.java
@@ -26,8 +26,8 @@ import org.destinationsol.common.SolRandom;
 import org.destinationsol.game.DebugOptions;
 import org.destinationsol.game.SaveManager;
 import org.destinationsol.game.SolGame;
-import org.destinationsol.game.context.Context;
-import org.destinationsol.game.context.internal.ContextImpl;
+import org.destinationsol.context.Context;
+import org.destinationsol.context.internal.ContextImpl;
 import org.destinationsol.game.sound.OggMusicManager;
 import org.destinationsol.game.sound.OggSoundManager;
 import org.destinationsol.menu.MenuScreens;
