+++ engine/src/main/java/org/destinationsol/context/internal/ContextImpl.java
@@ -1,0 +1,40 @@
+/*
+ * Copyright 2018 MovingBlocks
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package org.destinationsol.context.internal;
+
+import com.google.common.collect.Maps;
+import org.destinationsol.context.Context;
+
+import java.util.Map;
+
+public class ContextImpl implements Context {
+
+    private final Map<Class<?>, Object> map = Maps.newConcurrentMap();
+
+    @Override
+    public <T> T get(Class<? extends T> type) {
+        if (type == Context.class) {
+            return type.cast(this);
+        }
+        return type.cast(map.get(type));
+    }
+
+    @Override
+    public <T, U extends T> void put(Class<T> type, U object)  {
+        map.put(type, object);
+    }
+
+}
